+++ sdk/astro/arm-astro/CHANGELOG.md
@@ -1,15 +1,49 @@
 # Release History
-
-## 1.0.0-beta.2 (Unreleased)
-
+    
+## 1.0.0-beta.2 (2024-11-26)
+Compared with version 1.0.0-beta.1
+    
 ### Features Added
+
+  - Added operation Organizations.getResources
+  - Added operation Organizations.getRoles
+  - Added operation Organizations.getUsers
+  - Added operation Organizations.manageRoles
+  - Added operation Organizations.removeUser
+  - Added Interface GetResourcesRequest
+  - Added Interface GetResourcesSuccessResponse
+  - Added Interface GetRolesRequest
+  - Added Interface GetRolesSuccessResponse
+  - Added Interface GetUsersRequest
+  - Added Interface GetUsersSuccessResponse
+  - Added Interface LiftrBaseMarketplaceDetailsUpdate
+  - Added Interface LiftrBaseOfferDetailsUpdate
+  - Added Interface ManageRolesModel
+  - Added Interface OrganizationsGetResourcesOptionalParams
+  - Added Interface OrganizationsGetRolesOptionalParams
+  - Added Interface OrganizationsGetUsersOptionalParams
+  - Added Interface OrganizationsManageRolesOptionalParams
+  - Added Interface OrganizationsRemoveUserOptionalParams
+  - Added Interface PageInfo
+  - Added Interface PartnerResource
+  - Added Interface RemoveUserRequest
+  - Added Interface Role
+  - Added Interface User
+  - Added Type Alias OrganizationsGetResourcesResponse
+  - Added Type Alias OrganizationsGetRolesResponse
+  - Added Type Alias OrganizationsGetUsersResponse
+  - Added Type Alias OrganizationsManageRolesResponse
+  - Added Type Alias RenewalMode
+  - Interface LiftrBaseOfferDetails has a new optional parameter endDate
+  - Interface LiftrBaseOfferDetails has a new optional parameter renewalMode
+  - Interface OrganizationResourceUpdateProperties has a new optional parameter marketplace
+  - Added Enum KnownRenewalMode
 
 ### Breaking Changes
 
-### Bugs Fixed
-
-### Other Changes
-
+  - Removed Enum KnownVersions
+    
+    
 ## 1.0.0-beta.1 (2024-02-07)
 
 The package of @azure/arm-astro is using our next generation design principles. To learn more, please refer to our documentation [Quick Start](https://aka.ms/azsdk/js/mgmt/quickstart ).